chore: rename model filename
