+++ sinch/domains/numbers/models/v1/response/__init__.py
@@ -1,7 +1,7 @@
 from sinch.domains.numbers.models.v1.response.active_number import ActiveNumber
 from sinch.domains.numbers.models.v1.response.available_number import AvailableNumber
 from sinch.domains.numbers.models.v1.response.available_region import AvailableRegion
-from sinch.domains.numbers.models.v1.response.numbers_callback import CallbackConfigurationResponse
+from sinch.domains.numbers.models.v1.response.callback_configuration_response import CallbackConfigurationResponse
 
 __all__ = [
     "ActiveNumber",
